chore: remove the stream storage-layer row query egress
