+++ banyand/stream/tracing.go
@@ -18,18 +18,10 @@
 package stream
 
 import (
-	"context"
 	"fmt"
 	"time"
 
 	"github.com/dustin/go-humanize"
-
-	"github.com/apache/skywalking-banyandb/pkg/query"
-)
-
-const (
-	partMetadataHeader = "MinTimestamp, MaxTimestamp, CompressionSize, UncompressedSize, TotalCount, BlocksCount"
-	blockHeader        = "PartID, SeriesID, MinTimestamp, MaxTimestamp, Count, UncompressedSize"
 )
 
 func (pm *partMetadata) String() string {
@@ -48,31 +40,4 @@ func (bc *blockCursor) String() string {
 
 	return fmt.Sprintf("%d, %d, %s, %s, %d, %s",
 		bc.p.partMetadata.ID, bc.bm.seriesID, minTimestamp, maxTimestamp, bc.bm.count, humanize.Bytes(bc.bm.uncompressedSizeBytes))
-}
-
-func startBlockScanSpan(ctx context.Context, sids int, parts []*part, qr *idxResult) func() {
-	tracer := query.GetTracer(ctx)
-	if tracer == nil {
-		return func() {}
-	}
-
-	span, _ := tracer.StartSpan(ctx, "scan-blocks")
-	if qr.qo.elementFilter != nil {
-		span.Tag("filter_size", fmt.Sprintf("%d", qr.qo.elementFilter.Len()))
-	}
-	span.Tag("series_num", fmt.Sprintf("%d", sids))
-	span.Tag("part_header", partMetadataHeader)
-	span.Tag("part_num", fmt.Sprintf("%d", len(parts)))
-	for i := range parts {
-		span.Tag(fmt.Sprintf("part_%d_%s", parts[i].partMetadata.ID, parts[i].path),
-			parts[i].partMetadata.String())
-	}
-
-	return func() {
-		span.Tag("block_header", blockHeader)
-		for i := range qr.data {
-			span.Tag(fmt.Sprintf("block_%d", i), qr.data[i].String())
-		}
-		span.Stop()
-	}
 }
