+++ pkg/query/logical/stream/stream_plan_indexscan_local.go
@@ -18,20 +18,12 @@
 package stream
 
 import (
-	"context"
-	"encoding/hex"
 	"fmt"
-	"time"
-
-	"google.golang.org/protobuf/types/known/timestamppb"
 
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	streamv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/stream/v1"
-	"github.com/apache/skywalking-banyandb/pkg/convert"
 	"github.com/apache/skywalking-banyandb/pkg/index"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
-	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
 	"github.com/apache/skywalking-banyandb/pkg/query/executor"
 	"github.com/apache/skywalking-banyandb/pkg/query/logical"
 	"github.com/apache/skywalking-banyandb/pkg/query/model"
@@ -93,66 +85,4 @@ func (i *localIndexScan) Schema() logical.Schema {
 		return i.schema
 	}
 	return i.schema.ProjTags(i.projectionTagRefs...)
-}
-
-// BuildElementsFromStreamResult builds a slice of elements from the given stream query result.
-func BuildElementsFromStreamResult(ctx context.Context, result model.StreamQueryResult, projectionTags []model.TagProjection) (elements []*streamv1.Element, err error) {
-	var r *model.StreamResult
-	for {
-		r = result.Pull(ctx)
-		if r == nil {
-			return nil, nil
-		}
-		if r.Error != nil {
-			return nil, r.Error
-		}
-		if len(r.Timestamps) > 0 {
-			break
-		}
-	}
-	tagFamilyMap := make(map[string]*model.TagFamily, len(r.TagFamilies))
-	for idx := range r.TagFamilies {
-		tagFamilyMap[r.TagFamilies[idx].Name] = &r.TagFamilies[idx]
-	}
-	seenElementIDs := make(map[uint64]bool)
-	for i := range r.Timestamps {
-		elementID := r.ElementIDs[i]
-		// Deduplicate: skip if we've already seen this element ID
-		if seenElementIDs[elementID] {
-			continue
-		}
-		seenElementIDs[elementID] = true
-		e := &streamv1.Element{
-			Timestamp: timestamppb.New(time.Unix(0, r.Timestamps[i])),
-			ElementId: hex.EncodeToString(convert.Uint64ToBytes(elementID)),
-		}
-
-		for _, proj := range projectionTags {
-			tagFamily := &modelv1.TagFamily{
-				Name: proj.Family,
-			}
-			e.TagFamilies = append(e.TagFamilies, tagFamily)
-			resultTagFamily := tagFamilyMap[proj.Family]
-			for _, tagName := range proj.Names {
-				var tagValue *modelv1.TagValue
-				if resultTagFamily != nil {
-					for _, t := range resultTagFamily.Tags {
-						if t.Name == tagName {
-							tagValue = t.Values[i]
-							break
-						}
-					}
-				}
-				if tagValue == nil {
-					tagValue = pbv1.NullTagValue
-				}
-				tagFamily.Tags = append(tagFamily.Tags, &modelv1.Tag{
-					Key:   tagName,
-					Value: tagValue,
-				})
-			}
-		}
-		elements = append(elements, e)
-	}
-	return elements, nil
 }
